requirements.txt
Django==1.8.2
MySQL-python==1.2.3
django-bootstrap3==6.1.0
mysqlclient==1.3.6
oauthlib==0.7.2
python-social-auth==0.2.11
django-geoposition
urllib3
django-disqus
django-bootstrap3-datetimepicker